project-squad33-main/project-squad33-main/code/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
Authlib==1.2.0
Babel==2.10.3
blinker==1.5
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
coverage==6.2
cryptography==39.0.2
dataclasses==0.8
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.1
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-RESTful==0.3.9
flask-serialize==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
Flask-WhooshAlchemy==0.56
Flask-WTF==1.0.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.56.3
greenlet==1.1.3.post0
httplib2==0.21.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
infinity==1.5
intervals==0.9.2
itsdangerous==2.0.1
Jinja2==3.0.3
login==0.0.6
Mako==1.1.6
MarkupSafe==2.0.1
oauthlib==3.2.2
permissive-dict==1.0.4
phonenumbers==8.13.8
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.5
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.42
stripe==5.2.0
twilio==7.16.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
Werkzeug==2.0.3
Whoosh==2.7.4
WTForms==3.0.0
WTForms-Components==0.10.5
WTForms-Ext==0.5
WTForms-SQLAlchemy==0.3
zipp==3.6.0
